move: auth 폴더 위치 2차 변경 (#124)
