+++ apps/web/src/features/auth/components/SocialAuthSignUp.tsx
@@ -3,9 +3,9 @@ import { toast } from '@repo/ui/components/ui/sonner';
 import { FcGoogle } from 'react-icons/fc';
 import { IoChatbubble } from 'react-icons/io5';
 import { LuArrowRight } from 'react-icons/lu';
-import { selectSignUp } from 'src/entities/user/auth/supabase';
-import SocialAuthButton from 'src/features/user/auth/components/SocialAuthButton';
-import type { SocialAuthProvider } from 'src/entities/user/auth/types';
+import { selectSignUp } from 'src/entities/auth/supabase';
+import SocialAuthButton from 'src/features/auth/components/SocialAuthButton';
+import type { SocialAuthProvider } from 'src/entities/auth/types';
 
 const SocialAuthSignUp = () => {
   const handleSocialSignUp = async (provider: SocialAuthProvider) => {
